modified files

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Validator;
 use App\Item;
 use App\User;
@@ -35,25 +36,20 @@ class UserController extends Controller
         $reviews_count = count($reviews);
         $items = array();
         if($reviews_count > 0){
-            $items = Item::where(function ($query) use ($reviews) {
-                foreach ($reviews as $review) {
-                    $query->orWhere('items.id', $review->item_id);
-                }
-            })->take(20)->get();
+            $item = new Item();
+            $items = $item->getReviewItems($reviews);
         }
-        $followings_count = $user_detail->follows()->count();
-        $followers_count = Follow::where('follow_id', $user_detail->id)->count();
-        $follow = Follow::where([
-                                    ['user_id', '=', Auth::id()],
-                                    ['follow_id', '=', $user_detail->id]
-                                ])
-                              ->get();
+
+        $follow = new Follow();
+        $followings_count = $follow->getFollowingsCount($user_detail);
+        $followers_count = $follow->getFollowersCount($user_detail);
+        $my_follow = $follow->getMyFollow($user_detail->id);
 
         $follow_id = "";
         $follow_status = "";
-        if(count($follow) > 0){
+        if(count($my_follow) > 0){
             $follow_status = "active";
-            $follow_id = $follow[0]->id;
+            $follow_id = $my_follow[0]->id;
         }
         return view('user.index', [
             'user' => $user_detail,
@@ -76,6 +72,7 @@ class UserController extends Controller
     }
 
     public function store(Request $request){
+        \Log::info("UserController : store() : Start");
         $inputs = $request->all();
         $upload_file = $request->image;
         $name = $request->input('name');
@@ -84,16 +81,18 @@ class UserController extends Controller
 
         // ルールを設定
         $rules = [
+            'name' => 'required|string|max:20',
             'image' => 'image',
-            'name' => 'required|max:255',
-            'content' => 'max:300',
+            'content' => 'string|max:300',
         ];
 
         // エラーメッセージを設定
         $messages = [
-            'image.image' => '画像(jpg、png、bmp、gif、svg)であることを確認してください',
             'name.required' => 'ユーザー名は必須です',
-            'name.max' => 'ユーザー名は255文字以内です',
+            'name.string' => 'ユーザー名は文字入力です',
+            'name.max' => 'ユーザー名は20文字以内です',
+            'image.image' => '画像(jpg、png、bmp、gif、svg)であることを確認してください',
+            'content.string' => '自己紹介文は文字入力です',
             'content.max' => 'コメントは300文字以内です',
         ];
 
@@ -106,28 +105,39 @@ class UserController extends Controller
         $user = new User();
         $user_detail = $user->getUser($nickname);
 
-        $image = "";
-        if($upload_file == null){
-            $user_detail->name = $name;
-            $user_detail->nickname = $nickname;
-            $user_detail->content = $content;
+        DB::beginTransaction();
+        try {
+            $image = null;
+            if($upload_file == null){
+                $user_detail->name = $name;
+                $user_detail->nickname = $nickname;
+                $user_detail->content = $content;
 
-            $user_detail->save();
-        }else{
-            // 以前アバター画像を削除
-            \Storage::delete('public/images/users/'.$user_detail->image);
-            // アバター画像を保存
-            $image_encode = base64_encode($user_detail->nickname);
-            $image_name = str_replace(array('+','=','/'),array('_','-','.'),$image_encode);
-            $image = $image_name.'.jpg';
-            $upload_file->storeAs('/public/images/users', $image);
+                $user_detail->save();
+            }else{
+                if($user_detail->image != null){
+                    // 以前アバター画像を削除
+                    \Storage::delete('public/images/users/'.$user_detail->image);
+                }
+                // アバター画像を保存
+                $image_encode = base64_encode($user_detail->nickname);
+                $image_name = str_replace(array('+','=','/'),array('_','-','.'),$image_encode);
+                $image = $image_name.'.jpg';
+                $upload_file->storeAs('/public/images/users', $image);
 
-            $user_detail->name = $name;
-            $user_detail->nickname = $nickname;
-            $user_detail->content = $content;
-            $user_detail->image = $image;
+                $user_detail->name = $name;
+                $user_detail->nickname = $nickname;
+                $user_detail->content = $content;
+                $user_detail->image = $image;
 
-            $user_detail->save();
+                $user_detail->save();
+            }
+            DB::commit();
+        } catch (\Exception $e) {
+            DB::rollback();
+            \Log::emergency("UserController : store() : Failed Edit User! : user_id = ".$user_detail->id);
+            \Log::emergency("Message : ".$e->getMessage());
+            \Log::emergency("Code : ".$e->getCode());
         }
 
         return redirect()->route('user', ['nickname' => $nickname]);
